[spi_device] Add Last Read Address CSR
This commit adds last read address accessed by the host system. While
operating (SPI interface is in active), the CSR is not updated. Using
latch and CSb as an enable signal, the CSR is updated when the SPI
interface is in idle.

TODO: need to check if still 2FF sync on core clock domain is necessary
or not.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>